updated models and  implement functionality of  user registration and authentication with react login and employee registration form

+++ billing_market_backend/billing_market/stocks_app/models.py
@@ -12,9 +12,9 @@ class ProductCategory(models.Model):
 class GST(models.Model):
     gst_id = models.BigAutoField(primary_key=True)
     hsn_code = models.CharField(max_length=8, unique=True)
-    cgst = models.IntegerField(default=0)
-    sgst = models.IntegerField(default=0)
-    igst = models.IntegerField(default=0)
+    cgst = models.FloatField(default=0)
+    sgst = models.FloatField(default=0)
+    igst = models.FloatField(default=0)
     discription = models.TextField(blank=True)
 
     def __str__(self):
